Fixed for runs with openstack flag
When vsperf runs with the openstack flag, it fails because the 'connection' variable is not defined. This is solved by adding the 'connection' method from the openstack library.

After that, vsperf fails when finding the domain in the project. This happens because the 'OS_PROJECT_DOMAIN_ID' parameter is missing. I suggest to keep it simple and change OS_PROJECT_DOMAIN_ID to OS_PROJECT_DOMAIN_NAME in the conf file to run with the openstack flag, instead of adding this parameter to tools/os_deploy_tgen/utilities/utils.py.